BodyWeight listo

+++ My-Track-Fit-ui/app/src/main/java/com/example/my_track_fit/model/BodyWeight.kt
@@ -17,7 +17,7 @@ class BodyWeight(
     }
 
     //Methods
-    fun addBodyWeightMark(weight: Int) {
+    fun addBodyWeightMark(weight: Double) {
         val newMark = Mark(
             bodyWeightMark = weight,
             date = LocalDate.now()
